[IDLE-141] CenterEmployCardInfoView분리
